Move tree pointers to separate module
**Motivation**
The main `representation` module is quite large, and grouping all the
pointers together in a sub-module makes for a slightly easier-to-read
source code.

**Testing Done**
`./scripts/full-test.sh nightly`
